+++ library/src/main/java/com/getstream/sdk/chat/adapter/BaseMessageListItemViewHolder.kt
@@ -4,9 +4,6 @@ import android.content.Context
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.getstream.sdk.chat.view.MessageListView.BubbleHelper
-import com.getstream.sdk.chat.view.MessageListViewStyle
-import io.getstream.chat.android.client.models.Channel
 
 abstract class BaseMessageListItemViewHolder<T : MessageListItem>(
     resId: Int,
@@ -19,25 +16,18 @@ abstract class BaseMessageListItemViewHolder<T : MessageListItem>(
      * Workaround to allow a downcast of the MessageListItem to T
      */
     fun bindListItem(
-        context: Context,
-        channel: Channel,
         messageListItem: MessageListItem,
-        style: MessageListViewStyle,
-        bubbleHelper: BubbleHelper,
-        factory: MessageViewHolderFactory,
         position: Int
     ) {
         @Suppress("UNCHECKED_CAST")
-        bind(context, channel, messageListItem as T, style, bubbleHelper, factory, position)
+        bind(messageListItem as T, position)
     }
 
     protected abstract fun bind(
-        context: Context,
-        channel: Channel,
         messageListItem: T,
-        style: MessageListViewStyle,
-        bubbleHelper: BubbleHelper,
-        factory: MessageViewHolderFactory,
         position: Int
     )
+
+    protected val context: Context
+        get() = itemView.context
 }