+++ library/src/main/java/com/getstream/sdk/chat/adapter/MessageListItemViewHolder.java
@@ -1,7 +1,6 @@
 package com.getstream.sdk.chat.adapter;
 
 import android.annotation.SuppressLint;
-import android.content.Context;
 import android.graphics.Color;
 import android.graphics.drawable.Drawable;
 import android.text.SpannableString;
@@ -23,6 +22,7 @@ import androidx.annotation.IdRes;
 import androidx.annotation.NonNull;
 import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.constraintlayout.widget.ConstraintSet;
+import androidx.core.content.ContextCompat;
 import androidx.core.graphics.drawable.DrawableCompat;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
@@ -75,12 +75,12 @@ public class MessageListItemViewHolder extends BaseMessageListItemViewHolder<Mes
 
     protected RecyclerView.LayoutManager mLayoutManager;
 
-    protected Channel channel;
-    protected MessageListViewStyle style;
-    protected MessageListView.BubbleHelper bubbleHelper;
-    protected MessageViewHolderFactory viewHolderFactory;
+    protected final Channel channel;
+    protected final MessageListViewStyle style;
+    protected final MessageListView.BubbleHelper bubbleHelper;
+    protected final AttachmentViewHolderFactory viewHolderFactory;
+
     protected int position;
-    protected Context context;
     protected Message message;
     protected MessageListItem.MessageItem messageListItem;
 
@@ -105,7 +105,11 @@ public class MessageListItemViewHolder extends BaseMessageListItemViewHolder<Mes
 
     public MessageListItemViewHolder(
             int resId,
-            ViewGroup viewGroup,
+            @NonNull ViewGroup viewGroup,
+            @NonNull MessageListViewStyle style,
+            @NonNull Channel channel,
+            @NonNull AttachmentViewHolderFactory factory,
+            @NonNull MessageListView.BubbleHelper bubbleHelper,
             @NonNull MessageListView.MessageClickListener messageClickListener,
             @NonNull MessageListView.MessageLongClickListener messageLongClickListener,
             @NonNull MessageListView.AttachmentClickListener attachmentClickListener,
@@ -115,6 +119,10 @@ public class MessageListItemViewHolder extends BaseMessageListItemViewHolder<Mes
             @NonNull MessageListView.GiphySendListener giphySendListener
     ) {
         super(resId, viewGroup);
+        this.style = style;
+        this.channel = channel;
+        this.viewHolderFactory = factory;
+        this.bubbleHelper = bubbleHelper;
         this.messageClickListener = messageClickListener;
         this.messageLongClickListener = messageLongClickListener;
         this.attachmentClickListener = attachmentClickListener;
@@ -148,26 +156,14 @@ public class MessageListItemViewHolder extends BaseMessageListItemViewHolder<Mes
         pb_deliver = itemView.findViewById(R.id.pb_deliver);
         iv_deliver = itemView.findViewById(R.id.iv_deliver);
 
-        mLayoutManager = new LinearLayoutManager(context, RecyclerView.HORIZONTAL, false);
+        mLayoutManager = new LinearLayoutManager(getContext(), RecyclerView.HORIZONTAL, false);
         rv_reaction.setLayoutManager(mLayoutManager);
     }
 
     @Override
-    public void bind(@NonNull Context context,
-                     @NonNull Channel channel,
-                     @NonNull MessageListItem.MessageItem messageListItem,
-                     @NonNull MessageListViewStyle style,
-                     @NonNull MessageListView.BubbleHelper bubbleHelper,
-                     @NonNull MessageViewHolderFactory factory,
-                     int position) {
-
+    public void bind(@NonNull MessageListItem.MessageItem messageListItem, int position) {
         // set binding
-        this.context = context;
-        this.channel = channel;
         this.messageListItem = messageListItem;
-        this.style = style;
-        this.bubbleHelper = bubbleHelper;
-        this.viewHolderFactory = factory;
         this.position = position;
 
         this.message = messageListItem.getMessage();
@@ -352,7 +348,7 @@ public class MessageListItemViewHolder extends BaseMessageListItemViewHolder<Mes
             // Set Failed Message Title Text
             SpannableStringBuilder builder = new SpannableStringBuilder();
             int failedDes = TextUtils.isEmpty(message.getCommand()) ? R.string.stream_message_failed_send : R.string.stream_message_invalid_command;
-            SpannableString str1 = new SpannableString(context.getResources().getText(failedDes));
+            SpannableString str1 = new SpannableString(getContext().getResources().getText(failedDes));
             str1.setSpan(new ForegroundColorSpan(Color.GRAY), 0, str1.length(), 0);
             str1.setSpan(new RelativeSizeSpan(0.7f), 0, str1.length(), 0);
             builder.append(str1);
@@ -371,8 +367,8 @@ public class MessageListItemViewHolder extends BaseMessageListItemViewHolder<Mes
 
     protected void configMessageTextStyle() {
         if (isDeletedMessage()) {
-            tv_text.setTextSize(TypedValue.COMPLEX_UNIT_PX, context.getResources().getDimensionPixelSize(R.dimen.stream_message_deleted_text_font_size));
-            tv_text.setTextColor(context.getResources().getColor(R.color.stream_gray_dark));
+            tv_text.setTextSize(TypedValue.COMPLEX_UNIT_PX, getContext().getResources().getDimensionPixelSize(R.dimen.stream_message_deleted_text_font_size));
+            tv_text.setTextColor(getContext().getResources().getColor(R.color.stream_gray_dark));
             return;
         }
 
@@ -411,7 +407,7 @@ public class MessageListItemViewHolder extends BaseMessageListItemViewHolder<Mes
 
             if (isFailedMessage() && !ChatClient.instance().isSocketConnected())
                 return;
-            messageClickListener.onMessageClick(message, position);
+            messageClickListener.onMessageClick(message);
         });
 
         tv_text.setOnLongClickListener(view -> {
@@ -429,7 +425,7 @@ public class MessageListItemViewHolder extends BaseMessageListItemViewHolder<Mes
                     isLongClick = false;
                     return;
                 }
-                Chat.getInstance().getNavigator().navigate(new WebLinkDestination(url, context));
+                Chat.getInstance().getNavigator().navigate(new WebLinkDestination(url, getContext()));
             }
         });
     }
@@ -471,7 +467,7 @@ public class MessageListItemViewHolder extends BaseMessageListItemViewHolder<Mes
         rv_reaction.setVisibility(View.VISIBLE);
         iv_tail.setVisibility(View.VISIBLE);
         space_reaction_tail.setVisibility(View.VISIBLE);
-        rv_reaction.setAdapter(new ReactionListItemAdapter(context,
+        rv_reaction.setAdapter(new ReactionListItemAdapter(getContext(),
                 message.getReactionCounts(),
                 LlcMigrationUtils.getReactionTypes(),
                 style));
@@ -500,10 +496,10 @@ public class MessageListItemViewHolder extends BaseMessageListItemViewHolder<Mes
         tv_reply.setText(tv_reply.getContext().getResources().getQuantityString(R.plurals.stream_reply_count, replyCount, replyCount));
 
         iv_reply.setOnClickListener(view -> {
-            messageClickListener.onMessageClick(message, position);
+            messageClickListener.onMessageClick(message);
         });
         tv_reply.setOnClickListener(view -> {
-            messageClickListener.onMessageClick(message, position);
+            messageClickListener.onMessageClick(message);
         });
     }
 
@@ -625,7 +621,7 @@ public class MessageListItemViewHolder extends BaseMessageListItemViewHolder<Mes
                     params.endToEnd = R.id.space_reaction_tail;
             } else {
                 @DimenRes
-                int reactionMargin = context.getResources().getDimensionPixelSize(R.dimen.stream_reaction_margin);
+                int reactionMargin = getContext().getResources().getDimensionPixelSize(R.dimen.stream_reaction_margin);
                 if (tv_text.getWidth() + reactionMargin < rv_reaction.getWidth()) {
                     if (messageListItem.isMine())
                         params.endToEnd = R.id.tv_text;
@@ -648,7 +644,7 @@ public class MessageListItemViewHolder extends BaseMessageListItemViewHolder<Mes
     protected void configParamsUserAvatar() {
         if (avatar.getVisibility() != View.VISIBLE) return;
         ConstraintLayout.LayoutParams params = (ConstraintLayout.LayoutParams) avatar.getLayoutParams();
-        int marginStart = (int) context.getResources().getDimension(R.dimen.stream_message_avatar_margin);
+        int marginStart = (int) getContext().getResources().getDimension(R.dimen.stream_message_avatar_margin);
         if (messageListItem.isTheirs()) {
             params.startToStart = ConstraintLayout.LayoutParams.PARENT_ID;
             params.setMarginStart(marginStart);
@@ -733,18 +729,17 @@ public class MessageListItemViewHolder extends BaseMessageListItemViewHolder<Mes
                     .build());
 
             if (messageListItem.isMine())
-                iv_tail.setImageDrawable(context.getResources().getDrawable(R.drawable.stream_tail_outgoing));
+                iv_tail.setImageDrawable(ContextCompat.getDrawable(getContext(), R.drawable.stream_tail_outgoing));
             else
-                iv_tail.setImageDrawable(context.getResources().getDrawable(R.drawable.stream_tail_incoming));
+                iv_tail.setImageDrawable(ContextCompat.getDrawable(getContext(), R.drawable.stream_tail_incoming));
 
             DrawableCompat.setTint(iv_tail.getDrawable(), style.getReactionViewBgColor());
         } else {
             int drawable = style.getReactionViewBgDrawable();
-            rv_reaction.setBackground(context.getDrawable(drawable));
+            rv_reaction.setBackground(ContextCompat.getDrawable(getContext(), drawable));
             iv_tail.setVisibility(View.GONE);
         }
     }
-
 
     protected boolean isDeletedMessage() {
         return message.getDeletedAt() != null;