+++ library/src/main/java/com/getstream/sdk/chat/adapter/MessageViewHolderFactory.kt
@@ -6,10 +6,12 @@ import com.getstream.sdk.chat.adapter.MessageListItem.DateSeparatorItem
 import com.getstream.sdk.chat.adapter.MessageListItem.MessageItem
 import com.getstream.sdk.chat.adapter.MessageListItem.ThreadSeparatorItem
 import com.getstream.sdk.chat.adapter.MessageListItem.TypingItem
-import com.getstream.sdk.chat.model.ModelType
+import com.getstream.sdk.chat.view.MessageListView
+import com.getstream.sdk.chat.view.MessageListViewStyle
+import io.getstream.chat.android.client.models.Channel
 
 /**
- * Allows you to easily customize message rendering or message attachment rendering
+ * Allows you to easily customize message rendering
  */
 open class MessageViewHolderFactory {
     companion object {
@@ -18,14 +20,17 @@ open class MessageViewHolderFactory {
         const val MESSAGEITEM_TYPING = 3
         const val MESSAGEITEM_THREAD_SEPARATOR = 4
         const val MESSAGEITEM_NOT_FOUND = 5
-
-        const val GENERIC_ATTACHMENT = 1
-        const val IMAGE_ATTACHMENT = 2
-        const val VIDEO_ATTACHMENT = 3
-        const val FILE_ATTACHMENT = 4
     }
 
     lateinit var listenerContainer: ListenerContainer
+        @JvmName("setListenerContainerInternal")
+        internal set
+    lateinit var attachmentViewHolderFactory: AttachmentViewHolderFactory
+        @JvmName("setAttachmentViewHolderFactoryInternal")
+        internal set
+    lateinit var bubbleHelper: MessageListView.BubbleHelper
+        @JvmName("setBubbleHelperInternal")
+        internal set
 
     open fun getMessageViewType(messageListItem: MessageListItem?): Int {
         return when (messageListItem) {
@@ -37,34 +42,23 @@ open class MessageViewHolderFactory {
         }
     }
 
-    open fun getAttachmentViewType(
-        attachmentItem: AttachmentListItem
-    ): Int {
-        return when (attachmentItem.attachment.type) {
-            null ->
-                GENERIC_ATTACHMENT
-            ModelType.attach_video ->
-                VIDEO_ATTACHMENT
-            ModelType.attach_image, ModelType.attach_giphy ->
-                IMAGE_ATTACHMENT
-            ModelType.attach_file ->
-                FILE_ATTACHMENT
-            else ->
-                GENERIC_ATTACHMENT
-        }
-    }
-
     open fun createMessageViewHolder(
         parent: ViewGroup,
-        viewType: Int
+        viewType: Int,
+        style: MessageListViewStyle,
+        channel: Channel
     ): BaseMessageListItemViewHolder<*> {
         return when (viewType) {
             MESSAGEITEM_DATE_SEPARATOR ->
-                DateSeparatorViewHolder(R.layout.stream_item_date_separator, parent)
+                DateSeparatorViewHolder(R.layout.stream_item_date_separator, parent, style)
             MESSAGEITEM_MESSAGE ->
                 MessageListItemViewHolder(
                     R.layout.stream_item_message,
                     parent,
+                    style,
+                    channel,
+                    attachmentViewHolderFactory,
+                    bubbleHelper,
                     listenerContainer.messageClickListener,
                     listenerContainer.messageLongClickListener,
                     listenerContainer.attachmentClickListener,
@@ -74,30 +68,11 @@ open class MessageViewHolderFactory {
                     listenerContainer.giphySendListener
                 )
             MESSAGEITEM_TYPING ->
-                TypingIndicatorViewHolder(R.layout.stream_item_type_indicator, parent)
+                TypingIndicatorViewHolder(R.layout.stream_item_type_indicator, parent, style)
             MESSAGEITEM_THREAD_SEPARATOR ->
                 ThreadSeparatorViewHolder(R.layout.stream_item_thread_separator, parent)
             else ->
                 throw IllegalArgumentException("Unhandled viewType ($viewType)")
-        }
-    }
-
-    open fun createAttachmentViewHolder(
-        adapter: AttachmentListItemAdapter,
-        parent: ViewGroup,
-        viewType: Int
-    ): BaseAttachmentViewHolder {
-        return when (viewType) {
-            VIDEO_ATTACHMENT, IMAGE_ATTACHMENT ->
-                AttachmentViewHolderMedia(
-                    R.layout.stream_item_attach_media,
-                    parent,
-                    listenerContainer.giphySendListener
-                )
-            FILE_ATTACHMENT ->
-                AttachmentViewHolderFile(R.layout.stream_item_attachment_file, parent)
-            else ->
-                AttachmentViewHolder(R.layout.stream_item_attachment, parent)
         }
     }